[fix] moved jws extension
