+++ Sources/Entities/Validated/ValidatedSiopOpenId4VPRequest.swift
@@ -507,88 +507,6 @@ private extension ValidatedSiopOpenId4VPRequest {
   }
 }
 
-private extension JWS {
-    
-  // Function to convert Unix timestamp to Date
-  func dateFromUnixTimestamp(_ timestamp: Any) -> Date? {
-    if let timestampInt = timestamp as? Int {
-      return Date(timeIntervalSince1970: TimeInterval(timestampInt))
-    } else if let timestampDouble = timestamp as? Double {
-      return Date(timeIntervalSince1970: timestampDouble)
-    }
-    return nil
-  }
-    
-  func verifierAttestationClaims() throws -> VerifierAttestationClaims {
-      
-    let payload = payload.data()
-    guard let json = try JSONSerialization.jsonObject(
-        with: payload,
-        options: []
-    ) as? [String: Any] else {
-      throw ValidatedAuthorizationError.validationError("Invalid JWS payload")
-    }
-      
-    guard
-      let cnf = json["cnf"] as? [String: Any],
-      let jwkDict = cnf["jwk"] as? [String: Any],
-      let jwk = convertJSONToPublicKey(json: jwkDict)
-    else {
-      throw ValidatedAuthorizationError.validationError("Cannot locate cnf/jwk in payload")
-    }
-      
-    return VerifierAttestationClaims(
-      iss: try tryExtract(JWTClaimNames.issuer, from: json),
-      sub: try tryExtract(JWTClaimNames.subject, from: json),
-      iat: try tryExtract(JWTClaimNames.issuedAt, from: json, converter: dateFromUnixTimestamp),
-      exp: try tryExtract(JWTClaimNames.expirationTime, from: json, converter: dateFromUnixTimestamp),
-      verifierPubJwk: jwk,
-      redirectUris: try tryExtract("redirect_uris", from: json),
-      responseUris: try tryExtract("response_uris", from: json)
-    )
-  }
-    
-  // Function to convert JSON to ECPublicKey or RSAPublicKey
-  func convertJSONToPublicKey(json: [String: Any]) -> JWK? {
-    guard let kty = json["kty"] as? String else {
-      return nil
-    }
-        
-    switch kty {
-    case "EC":
-      return convertJSONToECPublicKey(json: json)
-    case "RSA":
-      return convertJSONToRSAPublicKey(json: json)
-    default:
-      return nil
-    }
-  }
-
-  // Function to convert JSON to ECPublicKey
-  func convertJSONToECPublicKey(json: [String: Any]) -> ECPublicKey? {
-    guard
-      let x = json["x"] as? String,
-      let y = json["y"] as? String,
-      let crv = json["crv"] as? String,
-      let curve = ECCurveType(rawValue: crv)
-    else {
-      return nil
-    }
-    return ECPublicKey(crv: curve, x: x, y: y)
-  }
-
-  // Function to convert JSON to RSAPublicKey
-  func convertJSONToRSAPublicKey(json: [String: Any]) -> RSAPublicKey? {
-    guard
-      let n = json["n"] as? String,
-      let e = json["e"] as? String
-    else {
-      return nil
-    }
-    return RSAPublicKey(modulus: n, exponent: e)
-  }
-}
-
 // Protocol to verify JWT claims
 private protocol JWTClaimsSetVerifier {
   func verify(claimsSet: JWTClaimsSet) throws
